Move internal/converter/json to internal/uimodel/converter/json
