+++ cmd/anonymizer/app/anonymizer/anonymizer.go
@@ -16,8 +16,8 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger-idl/model/v1"
-	uiconv "github.com/jaegertracing/jaeger/internal/converter/json"
 	"github.com/jaegertracing/jaeger/internal/uimodel"
+	uiconv "github.com/jaegertracing/jaeger/internal/uimodel/converter/json"
 )
 
 var allowedTags = map[string]bool{
