+++ cmd/query/app/http_handler.go
@@ -24,13 +24,13 @@ import (
 	"github.com/jaegertracing/jaeger-idl/model/v1"
 	"github.com/jaegertracing/jaeger/cmd/query/app/qualitymetrics"
 	"github.com/jaegertracing/jaeger/cmd/query/app/querysvc"
-	uiconv "github.com/jaegertracing/jaeger/internal/converter/json"
 	"github.com/jaegertracing/jaeger/internal/jtracer"
 	"github.com/jaegertracing/jaeger/internal/proto-gen/api_v2/metrics"
 	"github.com/jaegertracing/jaeger/internal/storage/metricstore/disabled"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/metricstore"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore"
 	ui "github.com/jaegertracing/jaeger/internal/uimodel"
+	uiconv "github.com/jaegertracing/jaeger/internal/uimodel/converter/json"
 )
 
 const (
