+++ crossdock/services/agent.go
@@ -13,7 +13,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger-idl/proto-gen/api_v2"
-	p2json "github.com/jaegertracing/jaeger/internal/converter/json"
+	p2json "github.com/jaegertracing/jaeger/internal/uimodel/converter/json"
 )
 
 var errSamplingRateMissing = errors.New("sampling rate is missing")
