+++ internal/sampling/http/handler.go
@@ -15,9 +15,9 @@ import (
 	"github.com/gorilla/mux"
 
 	"github.com/jaegertracing/jaeger-idl/proto-gen/api_v2"
-	p2json "github.com/jaegertracing/jaeger/internal/converter/json"
 	t2p "github.com/jaegertracing/jaeger/internal/converter/thrift/jaeger"
 	"github.com/jaegertracing/jaeger/internal/metrics"
+	p2json "github.com/jaegertracing/jaeger/internal/uimodel/converter/json"
 )
 
 const mimeTypeApplicationJSON = "application/json"
