+++ internal/uimodel/converter/json/fixtures/domain_01.json
